L5: fp32 accumulation, the audit it implied, and the shim deleted.

The fix is two casts in fused_masked_knn_topk: fp16xfp16 is exact in fp32, so
only the reduction changes. PTX at D=128 goes from 8 add.f16 / 0 f32 to 0
add.f16, 8 add.f32 + 14 fma.rn.f32. Max abs error against an fp64 dot on
goodreads-like inputs: 0.027644 -> 3.2e-6. linr_v2 torch-vs-triton jaccard@100
0.998743 -> 0.999751, and all 124 residual rows are exact fp16 ties on the
torch side, broken by torch.topk. Cost: B=1 0.05315 -> 0.05302 ms, B=16 0.9631
-> 0.9404 ms (2.4% faster).

The audit read every tl.sum / tl.dot / tl.reduce / cumsum in ops/triton/: the
fixed reduction was the ONLY floating-point reduction in the tree. The int8
scorers use tl.dot(out_dtype=int32), exact below 2^24; oporp sums int32
popcounts; the compaction reductions are integer. So no SilverTorch number
moves, and B3's head-to-head is unaffected.

The deliverable that outlasts the fix is tests/parity/test_accumulation.py: 8
tests bounding each scoring kernel against an fp64 computation of the same op,
each also asserting that an fp16 model of the same computation MISSES the
bound -- so the test's discriminating power is checked on every run. The old
parity suite could not see this class at all: it compares ops.triton against
ops.reference at the same input dtype, so both sides carried the same error and
agreed.

WP-2: the retrieve.layers / retrieve.kernels shim is deleted. It was temporary
tooling so the frozen golden worktree could run against the new library; the
golden stopped being a gate on 2026-09-15, so its reason to exist ended.

Library suite 653 passed (645 + 8), re-run by the orchestrator. Record:
linr-v2-backend-parity.md §8.1.

+++ retrieve/src/retrieve/ops/triton/fused_masked_knn_topk.py
@@ -16,8 +16,7 @@ from torch.library import triton_op, wrap_triton
 
 _P_BUCKETS = (256, 2048, 16384, 131072, 1048576)
 
-# Dtypes the kernel is exercised with today: parity tests feed fp32, PrefilterKNN feeds fp16.
-# Scores are always fp32 (the output buffer's dtype). Documents reality per kernels.md → I/O.
+# Parity tests feed fp32, PrefilterKNN feeds fp16; the dot accumulates in fp32 either way.
 _SUPPORTED_DTYPES = (torch.float16, torch.float32)
 
 
@@ -74,7 +73,9 @@ def _fused_masked_knn_topk_kernel(
     count = tl.load(counts_ptr + bid)
     in_count = n_offsets < count
 
-    q = tl.load(query_ptr + bid * stride_qb + d_offsets * stride_qd)
+    # Widen before the multiply: tl.sum reduces in its operand dtype (fp16 in production, plan
+    # L4 §6.1), and fp16 × fp16 is exact in fp32.
+    q = tl.load(query_ptr + bid * stride_qb + d_offsets * stride_qd).to(tl.float32)
 
     item_ids = tl.load(
         pos_indices_ptr + bid * stride_pb + n_offsets * stride_pp,
@@ -86,7 +87,7 @@ def _fused_masked_knn_topk_kernel(
         item_embs_ptr + item_ids[:, None] * stride_in + d_offsets[None, :] * stride_id,
         mask=in_count[:, None],
         other=0.0,
-    )
+    ).to(tl.float32)
 
     dots = tl.sum(emb_rows * q[None, :], axis=1)
     dots = tl.where(in_count, dots, float("-inf"))
